# docs/swagger.yaml
basePath: /
definitions:
  handlers.ApprovalRequest:
    properties:
      action:
        enum:
        - approve
        - reject
        type: string
      rental_id:
        type: integer
    required:
    - action
    - rental_id
    type: object
  handlers.BookingRequest:
    properties:
      airport_transfer:
        description: Optional
        type: boolean
      car_id:
        type: integer
      concierge_services:
        description: Optional
        type: boolean
      driver_id:
        type: integer
      dropoff_location:
        description: Optional
        type: string
      package_id:
        type: integer
      pickup_location:
        description: Optional
        type: string
      rental_date:
        type: string
      rental_duration:
        enum:
        - daily
        - weekly
        - monthly
        type: string
      return_date:
        type: string
    required:
    - car_id
    - driver_id
    - package_id
    - rental_date
    - rental_duration
    - return_date
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is Jakarta Luxury Rent Car service API documentation.
  title: Jakarta Luxury Rent Car API
  version: "1.0"
paths:
...
  /owner/approve-booking:
    post:
      consumes:
      - application/json
      description: Approve or reject a car booking
      parameters:
      - description: Approval request body containing rental ID and action (approve/reject)
        in: body
        name: approvalReq
        required: true
        schema:
          $ref: '#/definitions/handlers.ApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message indicating the booking has been approved or
            rejected
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format, validation error, car out of stock,
            or invalid action
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied. Only owners can approve or reject bookings.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User, car, or rental history not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update car stock, rental history, or send notifications
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve or reject a car booking
      tags:
      - Role Owner
...
  /users/booking:
    post:
      consumes:
      - application/json
      description: Book a car
      parameters:
      - description: Booking request body containing car ID and other booking details
        in: body
        name: bookingReq
        required: true
        schema:
          $ref: '#/definitions/handlers.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message and details of the car booking
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format, validation error, or insufficient stock
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Car, driver, or package not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create rental history, send notifications, or process
            booking
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Book a car
      tags:
      - Role User
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"